requirements.txt
numpy==1.21.6
python-dotenv==0.21.0
requests==2.28.1
terminaltables==3.1.10

